+++ src/client/java/com/culiatum/rp/client/CuliatumRpClient.java
@@ -1,0 +1,13 @@
+package com.culiatum.rp.client;
+
+import net.fabricmc.api.ClientModInitializer;
+import net.fabricmc.fabric.api.client.event.lifecycle.v1.ClientTickEvents;
+import net.fabricmc.fabric.api.client.rendering.v1.HudRenderCallback;
+
+public final class CuliatumRpClient implements ClientModInitializer {
+	@Override
+	public void onInitializeClient() {
+		ClientTickEvents.END_CLIENT_TICK.register(HandcuffsHud::tick);
+		HudRenderCallback.EVENT.register(HandcuffsHud::render);
+	}
+}